fix(review): full auto-review v0.5.7.0 — 보안 수정 + 설계 부채 등록
7명 리뷰어 병렬 디스패치(systems/logic/design/test/infra-security/docs-spec/docs-records)
전체 코드베이스 275 소스 + 193 문서 리뷰 → CRITICAL 4, MAJOR 26, MINOR 42건 발견.

코드 수정 (6건):
- [CRITICAL] chat_svc.toml Redis 포트 교차 수정 (6381→6382, 6382→6383)
- [MAJOR] Gateway 시스템 메시지 FlatBuffers Verifier 추가 (buffer overread 방어)
- [MAJOR] system_msg_ids uint16_t→uint32_t 통일 (WireHeader msg_id 타입 일치)
- [MAJOR] SlidingWindowCounter zero division 방어
- [MAJOR] ResponseDispatcher payload offset 경계 검증
- [MINOR] CI path filter cmake/** 추가

인프라 수정: 버전 번호 5곳 통일 (CMake/vcpkg → v0.5.7)
문서 수정: BACKLOG_HISTORY ID 보정 15건, 백로그 신규 9건(BACKLOG-89~97), 타이포 수정
빌드: MSVC /W4 /WX 성공, 71/71 유닛 통과

+++ apex_services/gateway/include/apex/gateway/gateway_service.hpp
@@ -46,9 +46,9 @@ class BroadcastFanout;
 /// 프레임워크가 직접 처리하는 시스템 메시지용이며, 서비스 라우팅을 거치지 않는다.
 struct system_msg_ids
 {
-    static constexpr uint16_t AUTHENTICATE_SESSION = 3;
-    static constexpr uint16_t SUBSCRIBE_CHANNEL = 4;
-    static constexpr uint16_t UNSUBSCRIBE_CHANNEL = 5;
+    static constexpr uint32_t AUTHENTICATE_SESSION = 3;
+    static constexpr uint32_t SUBSCRIBE_CHANNEL = 4;
+    static constexpr uint32_t UNSUBSCRIBE_CHANNEL = 5;
 };
 
 /// cross-core 글로벌 객체 — Server.global<T>()가 소유, 서비스는 raw pointer로 참조.
